# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Enterprise-QA-Agent API
  description: 混合检索 + LangGraph 多 Agent 企业问答（契约 docs/api-contract.md）
  version: 0.1.0
paths:
  /api/v1/chat:
    post:
      tags:
      - chat
      summary: Chat
      operationId: chat_api_v1_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents:
    post:
      tags:
      - documents
      summary: Upload Document
      operationId: upload_document_api_v1_documents_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_api_v1_documents_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/documents/{doc_id}:
    delete:
      tags:
      - documents
      summary: Delete Document
      description: F5.7 软删除：置全部有效 chunk is_valid=false（F1.7），不物理删除；重复删除幂等 204。
      operationId: delete_document_api_v1_documents__doc_id__delete
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
          title: Doc Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/tasks/{task_id}:
    get:
      tags:
      - tasks
      summary: Task Status
      operationId: task_status_api_v1_tasks__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/threads/{thread_id}/history:
    get:
      tags:
      - threads
      summary: Thread History
      operationId: thread_history_api_v1_threads__thread_id__history_get
      parameters:
      - name: thread_id
        in: path
        required: true
        schema:
          type: string
          title: Thread Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/debug/retrieve:
    post:
      tags:
      - debug
      summary: Debug Retrieve
      operationId: debug_retrieve_api_v1_debug_retrieve_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebugRetrieveRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/health:
    get:
      tags:
      - health
      summary: Health
      operationId: health_api_v1_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    Body_upload_document_api_v1_documents_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        meta:
          anyOf:
          - type: string
          - type: 'null'
          title: Meta
      type: object
      required:
      - file
      title: Body_upload_document_api_v1_documents_post
    ChatRequest:
      properties:
        thread_id:
          type: string
          maxLength: 128
          minLength: 1
          title: Thread Id
          description: checkpointer key 维度，建议 {tenant}:{user}
        question:
          type: string
          maxLength: 4000
          minLength: 1
          title: Question
        stream:
          type: boolean
          title: Stream
          default: true
        include_expired:
          type: boolean
          title: Include Expired
          default: false
      type: object
      required:
      - thread_id
      - question
      title: ChatRequest
    DebugRetrieveRequest:
      properties:
        query:
          type: string
          maxLength: 4000
          minLength: 1
          title: Query
        include_expired:
          type: boolean
          title: Include Expired
          default: false
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        top_k:
          type: integer
          maximum: 50.0
          minimum: 1.0
          title: Top K
          default: 8
      type: object
      required:
      - query
      title: DebugRetrieveRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError